chore: refactor it test to organize test domains in separate classes.
