+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/subclasses/descendants/testdomains/EntityB.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.spring.data.datastore.it.subclasses.descendants.testdomains;
+
+import com.google.cloud.spring.data.datastore.core.mapping.Descendants;
+import com.google.cloud.spring.data.datastore.core.mapping.DiscriminatorValue;
+import com.google.cloud.spring.data.datastore.core.mapping.Entity;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity(name = "A")
+@DiscriminatorValue("B")
+public class EntityB extends EntityA {
+  @Descendants private List<EntityC> entitiesC = new ArrayList<>();
+
+  public void addEntityC(EntityC entityCdescendants) {
+    this.entitiesC.add(entityCdescendants);
+  }
+
+  public List<EntityC> getEntitiesC() {
+    return entitiesC;
+  }
+}
