+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/subclasses/descendants/testdomains/SubclassesDescendantsEntityArepository.java
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.cloud.spring.data.datastore.it.subclasses.descendants.testdomains;
+
+import com.google.cloud.datastore.Key;
+import com.google.cloud.spring.data.datastore.repository.DatastoreRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SubclassesDescendantsEntityArepository extends DatastoreRepository<EntityA, Key> {}
