+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/subclasses/references/SubclassesReferencesIntegrationTests.java
@@ -18,36 +18,28 @@ package com.google.cloud.spring.data.datastore.it.subclasses.references;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.google.cloud.datastore.Key;
 import com.google.cloud.spring.data.datastore.core.DatastoreTemplate;
-import com.google.cloud.spring.data.datastore.core.mapping.DiscriminatorField;
-import com.google.cloud.spring.data.datastore.core.mapping.DiscriminatorValue;
-import com.google.cloud.spring.data.datastore.core.mapping.Entity;
 import com.google.cloud.spring.data.datastore.it.AbstractDatastoreIntegrationTests;
 import com.google.cloud.spring.data.datastore.it.DatastoreIntegrationTestConfiguration;
-import com.google.cloud.spring.data.datastore.repository.DatastoreRepository;
+import com.google.cloud.spring.data.datastore.it.subclasses.references.testdomains.EntityA;
+import com.google.cloud.spring.data.datastore.it.subclasses.references.testdomains.EntityB;
+import com.google.cloud.spring.data.datastore.it.subclasses.references.testdomains.EntityC;
+import com.google.cloud.spring.data.datastore.it.subclasses.references.testdomains.SubclassesReferencesEntityArepository;
 import java.util.Arrays;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.Reference;
-import org.springframework.stereotype.Repository;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@Repository
-interface SubclassesReferencesEntityArepository extends DatastoreRepository<EntityA, Key> {}
 
 @EnabledIfSystemProperty(named = "it.datastore", matches = "true")
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes = {DatastoreIntegrationTestConfiguration.class})
 public class SubclassesReferencesIntegrationTests extends AbstractDatastoreIntegrationTests {
 
-  @Autowired
-  SubclassesReferencesEntityArepository entityArepository;
+  @Autowired SubclassesReferencesEntityArepository entityArepository;
 
   @Autowired private DatastoreTemplate datastoreTemplate;
 
@@ -65,33 +57,5 @@ public class SubclassesReferencesIntegrationTests extends AbstractDatastoreInteg
     entityArepository.saveAll(Arrays.asList(entityB1, entityC1));
     EntityC fetchedC = (EntityC) entityArepository.findById(entityC1.getId()).get();
     assertThat(fetchedC.getEntityB()).isNotNull();
-  }
-}
-
-@Entity(name = "A")
-@DiscriminatorField(field = "type")
-abstract class EntityA {
-  @Id private Key id;
-
-  public Key getId() {
-    return id;
-  }
-}
-
-@Entity(name = "A")
-@DiscriminatorValue("B")
-class EntityB extends EntityA {}
-
-@Entity(name = "A")
-@DiscriminatorValue("C")
-class EntityC extends EntityA {
-  @Reference private EntityB entityB;
-
-  EntityC(EntityB entityB) {
-    this.entityB = entityB;
-  }
-
-  public EntityB getEntityB() {
-    return entityB;
   }
 }
