+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/subclasses/references/testdomains/EntityB.java
@@ -1,0 +1,8 @@
+package com.google.cloud.spring.data.datastore.it.subclasses.references.testdomains;
+
+import com.google.cloud.spring.data.datastore.core.mapping.DiscriminatorValue;
+import com.google.cloud.spring.data.datastore.core.mapping.Entity;
+
+@Entity(name = "A")
+@DiscriminatorValue("B")
+public class EntityB extends EntityA {}
