+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/testdomains/DogRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2019 the original author or authors.
+ * Copyright 2023 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,15 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spring.data.datastore.it;
+package com.google.cloud.spring.data.datastore.it.testdomains;
 
-import com.google.cloud.datastore.Key;
-import com.google.cloud.spring.data.datastore.entities.Product;
 import com.google.cloud.spring.data.datastore.repository.DatastoreRepository;
-import org.springframework.stereotype.Repository;
+import com.google.cloud.spring.data.datastore.repository.query.Query;
+import java.util.List;
 
-@Repository
-public interface ProductRepository extends DatastoreRepository<Product, Key> {}
+public interface DogRepository extends DatastoreRepository<Dog, Long> {
+  List<Dog> findByName(String s);
+
+  @Query("select * from Pet")
+  List<Dog> findByCustomQuery();
+}
