+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/testdomains/LazyEntity.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2018 the original author or authors.
+ * Copyright 2017-2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,21 +14,32 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spring.data.datastore.entities;
+package com.google.cloud.spring.data.datastore.it.testdomains;
 
 import com.google.cloud.spring.data.datastore.core.mapping.Entity;
+import com.google.cloud.spring.data.datastore.core.mapping.LazyReference;
 import java.util.Objects;
 import org.springframework.data.annotation.Id;
 
 @Entity
-public class ServiceConfiguration {
-  @Id private String serviceName;
+public class LazyEntity {
+  @Id
+  public Long id;
 
-  private CustomMap customMap;
+  @LazyReference LazyEntity lazyChild;
 
-  public ServiceConfiguration(String serviceName, CustomMap customMap) {
-    this.serviceName = serviceName;
-    this.customMap = customMap;
+  public LazyEntity() {}
+
+  public LazyEntity(LazyEntity child) {
+    this.lazyChild = child;
+  }
+
+  Long getId() {
+    return this.id;
+  }
+
+  LazyEntity getLazyChild() {
+    return this.lazyChild;
   }
 
   @Override
@@ -39,13 +50,18 @@ public class ServiceConfiguration {
     if (o == null || getClass() != o.getClass()) {
       return false;
     }
-    ServiceConfiguration that = (ServiceConfiguration) o;
-    return Objects.equals(serviceName, that.serviceName)
-        && Objects.equals(customMap, that.customMap);
+    LazyEntity that = (LazyEntity) o;
+    return Objects.equals(getId(), that.getId())
+        && Objects.equals(getLazyChild(), that.getLazyChild());
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(serviceName, customMap);
+    return Objects.hash(getId(), getLazyChild());
+  }
+
+  @Override
+  public String toString() {
+    return "LazyEntity{" + "id=" + this.id + ", lazyChild=" + this.lazyChild + '}';
   }
 }
