+++ spring-cloud-gcp-data-datastore/src/test/java/com/google/cloud/spring/data/datastore/it/testdomains/PetRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2018 the original author or authors.
+ * Copyright 2023 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.cloud.spring.data.datastore.it;
+package com.google.cloud.spring.data.datastore.it.testdomains;
 
-/** A projection interface for the integration tests. */
-public interface TestEntityProjection {
+import com.google.cloud.spring.data.datastore.repository.DatastoreRepository;
+import java.util.List;
 
-  String getColor();
+public interface PetRepository extends DatastoreRepository<Pet, Long> {
+  List<Pet> findByName(String s);
 }
